Moved checking of database table to start-up of bot Moved all database related code to db service
